src/test/java/com/study/chapter7/TransmissionTest.java
package com.study.chapter7;

import org.junit.Rule;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.junit.jupiter.api.extension.RegisterExtension;
import org.junit.rules.ExpectedException;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.equalTo;

public class TransmissionTest {

    private Transmission transmission;
    private Car car;

    @BeforeEach
    public void setup() {
        car = new Car();
        transmission = new Transmission(car);
    }

    @Rule
    public ExpectedException exception = ExpectedException.none();

    @Test
    public void remainsInDriveAfterAcceleration() {
        transmission.shift(Gear.DRIVE);
        car.accelerateTo(35);

        assertThat(transmission.getGear(), equalTo(Gear.DRIVE));
    }

    @Test
    public void ignoresShiftToParkWhileInDrive() {
        transmission.shift(Gear.DRIVE);
        car.accelerateTo(30);

        transmission.shift(Gear.PARK);

        assertThat(transmission.getGear(), equalTo(Gear.DRIVE));
    }

    @DisplayName("side effects 검사")
    @Test
    public void allowShiftToParkWhenNotMoving() {
        transmission.shift(Gear.DRIVE);
        car.accelerateTo(30);
        car.brakeToStop();

        transmission.shift(Gear.PARK);

        assertThat(transmission.getGear(), equalTo(Gear.PARK));
    }

    @org.junit.Test
    public void notShiftWhenNotGear() {
        exception.expect(GearNullException.class);
        exception.expectMessage("Gear에 값이 없습니다");

        transmission.shift(null);
    }
}